requirements.txt
langchain>=0.0.267
pypdf>=3.15.1
python-dotenv>=1.0.0
openai>=1.3.0
chromadb>=0.4.13
tiktoken>=0.5.1
pycryptodome>=3.19.0
streamlit>=1.24.0